+++ src/types/accounting/currency.ts
@@ -1,0 +1,9 @@
+/**
+ * Available currencies
+ */
+export enum Currency {
+  /**
+   * USA dollars
+   */
+  USD = 'USD'
+}
